[feat] - cont. default template exec
