--- mta/root.go
@@ -1,11 +1,0 @@
-package mta
-
-// MTA - Main mta struct
-type MTA struct {
-	SchemaVersion *string      `yaml:"_schema-version"`
-	Id            string       `yaml:"ID"`
-	Version       string       `yaml:"version,omitempty"`
-	Modules       []*Modules   `yaml:"modules,omitempty"`
-	Resources     []*Resources `yaml:"resources,omitempty"`
-	Parameters    Parameters   `yaml:"parameters,omitempty"`
-}
